# dont use shutdownRef count for sync completion queue