Save AtomicCounter allocation for sync unary call (currently there's 1 alloc per call).
The optimization is possible because keeping track of unfinished cq operations is only useful for async completion queues.
At the same time, allocating AtomicCounter for async completion queues doesn't bother us as the async cqs only get created very rarely (at startup).

Supersedes
https://github.com/grpc/grpc/pull/19536