@jtattermusch 2019-07-24
Before
```
BenchmarkDotNet=v0.11.5, OS=debian 10
Intel Xeon CPU 2.30GHz, 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.1.503
  [Host] : .NET Core 2.1.7 (CoreCLR 4.6.27129.04, CoreFX 4.6.27129.04), 64bit RyuJIT
  Core   : .NET Core 2.1.7 (CoreCLR 4.6.27129.04, CoreFX 4.6.27129.04), 64bit RyuJIT

Job=Core  Runtime=Core  

|                Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
| SyncUnaryCallOverhead | 7.879 us | 0.1725 us | 0.5004 us | 0.1068 |     - |     - |     808 B |
```


After
```
|                Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
| SyncUnaryCallOverhead | 8.784 us | 0.1699 us | 0.2209 us | 0.0763 |     - |     - |     784 B |
```

Which shows the saved invocation of AtomicCounter.